feat: extract a storage manager behaviour
